e2e-tests/requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.7
pycparser==2.21
PySocks==1.7.1
pytest==7.4.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
tomli==2.0.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
urllib3==2.0.4
wsproto==1.2.0
